// open_contabilidad/src/interfaces/ILibroDiario.ts
interface LibroDiario {
    numerof: number;
    tipof: string;
    nodoclf: number;
    fechaf: Date;
    glosaf: string;
    valorf: number;
    tipo: string;
    numero: number;
    fecha: Date;
    cuenta: number;
    obra: string;
    ncuenta: string;
    auxiliar: string;
    td: string;
    nodoc: number;
    debe: number;
    haber: number;
    glosa: string;
    nroref: string;
    codant: string;
    nombreTipoDoc: string;
  }
  
  export default LibroDiario;